+++ source/InputManager.cpp
@@ -15,8 +15,8 @@ namespace Trn {
 		// InputManager *pInputManager = static_cast<InputManager*>(glfwGetWindowUserPointer(window));
 		if (action == GLFW_PRESS) {
 			input.previousState[key] = input.currentState[key];
-			float currentTime = glfwGetTime();
-			float timeDiff = currentTime - input.lastTimePressed[key];
+			double currentTime = glfwGetTime();
+			double timeDiff = currentTime - input.lastTimePressed[key];
 			// using std::cout;
 			// using std::endl;
 			if ((timeDiff < input.doublePressInterval) && input.previousState[key] == KeyState::Released) {
